+++ src/main/java/br/com/cefet/controller/VeiculoController.java
@@ -15,16 +15,23 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
 import org.springframework.web.servlet.ModelAndView;
 
 import br.com.cefet.dto.requisicaoVeiculo;
 import br.com.cefet.model.Categoria;
+import br.com.cefet.model.Cliente;
 import br.com.cefet.model.Filial;
+import br.com.cefet.model.Funcionario;
 import br.com.cefet.model.Marca;
 import br.com.cefet.model.Paletas;
 import br.com.cefet.model.Veiculo;
 import br.com.cefet.repository.FilialRepository;
 import br.com.cefet.repository.VeiculoRepository;
+import br.com.cefet.service.SessaoService;
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 
 @Controller
@@ -36,11 +43,21 @@ public class VeiculoController {
 
 	@Autowired
 	private FilialRepository filialRepository;
+	
+	@Autowired
+	private SessaoService sessaoService;
 
 	@GetMapping("/veiculos")
-	public ModelAndView index() {
-//		List<Veiculo> veiculos = this.veiculoRepository.findByStatus("Disponível");
-		List<Veiculo> veiculos = this.veiculoRepository.findByStatusAndEmManutencao("Disponível", false);
+	public ModelAndView index(@RequestParam(name = "marca", required = false) Marca marca, @RequestParam(name = "categoria", required = false) Categoria categoria) {
+//		List<Veiculo> veiculos = this.veiculoRepository.findByStatusAndEmManutencao("Disponível", false);
+		List<Veiculo> veiculos;
+		 if (marca != null) {
+		        veiculos = this.veiculoRepository.findByMarcaVeiculo(marca);
+		    } else if (categoria != null){
+		    	veiculos = this.veiculoRepository.findByCategoriaVeiculo(categoria);
+		    } else {
+		        veiculos = this.veiculoRepository.findByStatusAndEmManutencao("Disponível", false);
+		    }
 		ModelAndView mv = new ModelAndView("veiculos/index");
 		mv.addObject("veiculos", veiculos);
 
@@ -49,7 +66,13 @@ public class VeiculoController {
 
 	@GetMapping("/veiculos/new")
 	public ModelAndView novo() {
-
+		HttpSession session = ((ServletRequestAttributes) RequestContextHolder.currentRequestAttributes()).getRequest()
+				.getSession();
+		Funcionario funcionario = sessaoService.obterFuncionarioDaSessao(session);
+		if (funcionario == null) {
+			System.out.println("Acesso negado!");
+			return new ModelAndView("redirect:/veiculos");
+		}
 		ModelAndView mv = new ModelAndView("veiculos/new");
 		List<Filial> filiais = filialRepository.findAll();
 		mv.addObject("filiais", filiais);
@@ -65,7 +88,8 @@ public class VeiculoController {
 	public requisicaoVeiculo getRequisicaoVeiculo() {
 		return new requisicaoVeiculo();
 	}
-
+	
+	
 	@PostMapping("/veiculos")
 	public ModelAndView create(@ModelAttribute requisicaoVeiculo requisicao, BindingResult result) {
 		ModelAndView mv = new ModelAndView("/veiculos/new");
@@ -84,6 +108,16 @@ public class VeiculoController {
 				Veiculo veiculo = new Veiculo();
 				veiculo.setBranch(filial);
 				veiculo = requisicao.toVeiculo(veiculo);
+				List<Veiculo> placa = veiculoRepository.findByPlaca(veiculo.getPlaca());
+				if (!placa.isEmpty()) {
+					System.out.println("Placa já cadastrada!");
+					List<Filial> filiais = filialRepository.findAll();
+					mv.addObject("filiais", filiais);
+					mv.addObject("marcaVeiculo", Marca.values());
+					mv.addObject("categoria", Categoria.values());
+					mv.addObject("cor", Paletas.values());
+					return mv;
+				}
 				veiculo.setStatus("Disponível");
 				veiculo.setEmManutencao(false);
 
@@ -114,7 +148,13 @@ public class VeiculoController {
 	@GetMapping("/veiculos/{id}/edit")
 	public ModelAndView edit(@PathVariable Integer id, requisicaoVeiculo requisicao) {
 		Optional<Veiculo> optional = this.veiculoRepository.findById(id);
-
+		HttpSession session = ((ServletRequestAttributes) RequestContextHolder.currentRequestAttributes()).getRequest()
+				.getSession();
+		Funcionario funcionario = sessaoService.obterFuncionarioDaSessao(session);
+		if (funcionario == null) {
+			System.out.println("Acesso negado!");
+			return new ModelAndView("redirect:/veiculos");
+		}
 		if (optional.isPresent()) {
 			System.out.printf("%d", id);
 			Veiculo veiculo = optional.get();
@@ -162,6 +202,13 @@ public class VeiculoController {
 
 	@GetMapping("/veiculos/{id}/delete")
 	public String delete(@PathVariable Integer id) {
+		HttpSession session = ((ServletRequestAttributes) RequestContextHolder.currentRequestAttributes()).getRequest()
+				.getSession();
+		Funcionario funcionario = sessaoService.obterFuncionarioDaSessao(session);
+		if (funcionario == null) {
+			System.out.println("Acesso negado!");
+			return "redirect:/veiculos";
+		}
 		try {
 			this.veiculoRepository.deleteById(id);
 			return "redirect:/veiculos";
